+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Null reference exception on AnimatorStateMachine.IsExecuting function
+
 ## [2.1.0] - 2023-03-30
 ### Added
 - AnimatorStateMachine.Animator property